created widget file for sample list tile
